Fix armv7l-linux jobs.
Until just a few hours ago, upstream's 'release-lib.nix' did not
support armv7l-linux builds. It's fixed now, but as we're not on that
revision yet, we need to copy its release-lib.nix to this repo and
make a few changes to make it work locally.

Eventually we can undo this hack and use upstream.