Better Hydra visibility for aggregate jobs.
'releaseTools.aggregate' is a very nice, convenient way to define a
subset of jobs, but Hydra treats the aggregate as a single job, so if
you want to drill down and examine one of the constituents, it doesn't
have the same kind of metadata as a full Hydra job. This commit gives
us the best of both worlds.