Disable the ghc-aarch64 fork jobset; something really weird going on.
For some reason, when this job is *evaluated* by the Hydra, the
evaluation ends up building the *packages*. I have no idea why but the
result is an evaluation that lasts for several hours and eventually
dumps a massive error log.

Anyway, hopefully the patch to re-enable parallel builds of GHC on
aarch64 will soon be merged upstream and I won't need to special-case
these builds anymore.