Bump dependencies.
Note that the nixpkgs here is from my own fork, which fixes a few
Haskell-related aarch64 issues. I'll switch back to upstream once
they're incorporated.

Also note that the strange evaluation issues I was seeing with aarch64
builds in the previous commits was due to the fact that the pinpon,
hpio, and mellon repos all use callCabal2nix to generate their Nix
files, and Hydra was trying to build it for aarch64, since it didn't
yet exist (the NixOS Hydra hadn't successfully built it yet due to
aforementioned aarch64 Haskell issues).

This explains the extremely long (and continuously failing) evaluation
phases.