Replace unbound-block-hosts with StevenBlack hosts (badhosts).
Note that the new badhosts package builds an Unbound-compatible block
list using always_nxdomain for each domain name in the source hosts
file. According to
https://github.com/StevenBlack/hosts/issues/237#issuecomment-348050199,
for recent versions of Unbound, this should be the fastest method.

Also note that, currently, the badhosts package only builds the base
"unified" hosts list, but not any of the additional filters.

Finally, this commit also removes the special-case `blockList` option
in the unbound-multi-instance NixOS module. To include a block list,
you can just add the badhosts package's unbound.conf file to
unbound-multi-instance's extraConfig option.