ssh default config: don't allow agent forwarding.
Yes, the sshd_config(5) man page points out that this doesn't prevent
users from installing their own forwaders. This isn't meant to prevent
that scenario. It's meant to prevent accidents; e.g., where the user
is in the habit of adding a "-A" flag to ssh, or perhaps they enabled
agent forwarding in their .ssh/config and forgot about it.