Add minikube and friends.
This uses a more recent version than upstream nixpkgs. It also builds
docker-machine-hyperkit separately from minikube.

This is based on Atkins Chang's pull request:
https://github.com/NixOS/nixpkgs/pull/66578

Note: the docker-machine-hyperkit minikube plugin is currently
disabled, because minikube on macOS needs to run it as root via
setuid, and this is incompatible with the Nix store. I don't know of a
way to wrap this appropriately on nix-darwin yet. At any rate,
minikube will download a compatible version automatically, and it
should work as long as hyperkit is in the minikube path, which this
minikube derivation does on macOS.